requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
pillow
pytesseract
